Assignment week 2
In this assignment, we’ll practice working with one-to-many relationships in SQL. Suppose you are tasked with keeping
track of a database that contain the best “how-to” videos on MySQL.
You may want to first create a new database (schema) for this assignment.
1. Videos table. Create one table to keep track of the videos. This table should include a unique ID, the title of the
video, the length in minutes, and the URL. Populate the table with at least three related videos from YouTube or
other publicly available resources.
2. Create and populate Reviewers table. Create a second table that provides at least two user reviews for each of
at least two of the videos. These should be imaginary reviews that include columns for the user’s name
(“Asher”, “Cyd”, etc.), the rating (which could be NULL, or a number between 0 and 5), and a short text review
(“Loved it!”). There should be a column that links back to the ID column in the table of videos.
3. Report on Video Reviews. Write a JOIN statement that shows information from both tables.